requirements.txt
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
python-dotenv>=1.0.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
pydantic>=2.5.2
fastapi>=0.104.1
uvicorn>=0.24.0
